move StreamTaskContext to its own class to avoid muzzle issues
