+++ dd-java-agent/instrumentation/kafka-streams-0.11/src/latestDepTest/groovy/KafkaStreamsTest.groovy
@@ -41,7 +41,7 @@ class KafkaStreamsTest extends AgentTestRunner {
 
   @Override
   protected boolean isDataStreamsEnabled() {
-    return true;
+    return true
   }
 
   def "test kafka produce and consume with streams in-between"() {
